requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
boto3==1.26.13
botocore==1.29.13
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap5==22.1
django-extensions==3.2.1
django-imagekit==4.1.0
django-multiselectfield==0.1.12
django-storages==1.13.1
executing==1.1.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.3.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.5.0
tzdata==2022.6
urllib3==1.26.12
wcwidth==0.2.5
